refac: created booster.ts and poolAndVaults.ts helpers
